requirements.txt
streamlit
huggingface_hub
openai>=1.0.0
python-docx
fpdf
PyMuPDF